Pulled netConn and Writer into separate files.
